requirements.dev.txt
appnope==0.1.2
backcall==0.2.0
bcrypt==3.2.0
certifi==2020.12.5
cffi==1.14.5
click==7.1.2
decorator==4.4.2
dnspython==2.1.0
email-validator==1.1.2
fastapi==0.63.0
h11==0.12.0
idna==3.1
ipython==7.21.0
ipython-genutils==0.2.0
jedi==0.18.0
mysql-connector-python==8.0.23
parso==0.8.1
pexpect==4.8.0
pickleshare==0.7.5
prompt-toolkit==3.0.16
protobuf==3.15.4
ptyprocess==0.7.0
pycparser==2.20
pydantic==1.8
Pygments==2.8.0
PyMySQL==1.0.2
pyodbc==4.0.30
python-dotenv==0.15.0
six==1.15.0
SQLAlchemy==1.3.23
starlette==0.13.6
traitlets==5.0.5
typing-extensions==3.7.4.3
uvicorn==0.13.4
wcwidth==0.2.5
